requirements.txt
streamlit>=1.24.0
easyocr>=1.7.0
numpy>=1.24.3
Pillow>=9.5.0
torch>=2.0.1
